+++ packages/core/src/index.ts
@@ -26,6 +26,8 @@
 /// <reference path="./types/geokdbush.d.ts" />
 /// <reference path="./types/geoflatbush.d.ts" />
 
+export type { IdOrIndex } from "./ids"
+export * from "./mocks"
 export * from "./nodes"
 export * from "./osm"
 export * from "./relations"