+++ packages/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @osmix/core
 
+## 0.1.7
+
+### Patch Changes
+
+- 2a634cb: Fix publishing
+- Updated dependencies [2a634cb]
+  - @osmix/shared@0.0.12
+
+## 0.1.6
+
+### Patch Changes
+
+- 3c8ee95: Fix and simplify package exports
+- Updated dependencies [3c8ee95]
+  - @osmix/shared@0.0.11
+
 ## 0.1.5
 
 ### Patch Changes