+++ packages/pbf/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osmix/pbf
 
+## 0.0.8
+
+### Patch Changes
+
+- 2a634cb: Fix publishing
+
+## 0.0.7
+
+### Patch Changes
+
+- 3c8ee95: Fix and simplify package exports
+
 ## 0.0.6
 
 ### Patch Changes