+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osmix/shared
 
+## 0.0.12
+
+### Patch Changes
+
+- 2a634cb: Fix publishing
+
+## 0.0.11
+
+### Patch Changes
+
+- 3c8ee95: Fix and simplify package exports
+
 ## 0.0.10
 
 ### Patch Changes