+++ packages/shortbread/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @osmix/shortbread
 
+## 0.0.10
+
+### Patch Changes
+
+- 2a634cb: Fix publishing
+- Updated dependencies [2a634cb]
+  - @osmix/shared@0.0.12
+  - @osmix/core@0.1.7
+  - @osmix/vt@0.0.12
+
+## 0.0.9
+
+### Patch Changes
+
+- 3c8ee95: Fix and simplify package exports
+- Updated dependencies [3c8ee95]
+  - @osmix/core@0.1.6
+  - @osmix/shared@0.0.11
+  - @osmix/vt@0.0.11
+
 ## 0.0.8
 
 ### Patch Changes