+++ packages/vt/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @osmix/vt
 
+## 0.0.12
+
+### Patch Changes
+
+- 2a634cb: Fix publishing
+- Updated dependencies [2a634cb]
+  - @osmix/shared@0.0.12
+
+## 0.0.11
+
+### Patch Changes
+
+- 3c8ee95: Fix and simplify package exports
+- Updated dependencies [3c8ee95]
+  - @osmix/shared@0.0.11
+
 ## 0.0.10
 
 ### Patch Changes